src/test/java/com/myapp/bookservice/service/impl/BookServiceImplTest.java
package com.myapp.bookservice.service.impl;

import com.myapp.bookservice.model.BookDTO;
import com.myapp.bookservice.repos.BookRepositoryImpl;
import com.myapp.bookservice.util.StringUtils;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.when;

public class BookServiceImplTest {

    @Test
    public void testGetAllBooksSortedByTitleDescending() {
        // Arrange
        BookRepositoryImpl bookRepository = Mockito.mock(BookRepositoryImpl.class);
        BookServiceImpl bookService = new BookServiceImpl(bookRepository);
        List<BookDTO> expectedBooks = new ArrayList<>();
        // Add some dummy data to expectedBooks

        // Mocking bookRepository behavior for the method call
        when(bookRepository.findAllByOrderByTitleDesc()).thenReturn(expectedBooks);

        // Act
        List<BookDTO> actualBooks = bookService.getAllBooksSortedByTitleDescending();

        // Assert
        assertEquals(expectedBooks, actualBooks);
    }

    @Test
    public void testGetBooksGroupedByAuthor() {
        // Arrange
        BookRepositoryImpl bookRepository = Mockito.mock(BookRepositoryImpl.class);
        BookServiceImpl bookService = new BookServiceImpl(bookRepository);
        List<BookDTO> allBooks = new ArrayList<>();
        // Add some dummy data to allBooks

        // Mocking bookRepository behavior for the method call
        when(bookRepository.findAll()).thenReturn(allBooks);

        // Act
        Map<String, List<BookDTO>> booksGroupedByAuthor = bookService.getBooksGroupedByAuthor();

        // Assert
        // You can add assertions to verify the grouping logic
    }

    @Test
    public void testCreate() {
        // Arrange
        BookRepositoryImpl bookRepository = Mockito.mock(BookRepositoryImpl.class);
        BookServiceImpl bookService = new BookServiceImpl(bookRepository);
        BookDTO bookDTO = new BookDTO();
        // Set bookDTO properties

        Long expectedId = 1L;
        // Mocking behavior of bookRepository.save
        when(bookRepository.save(bookDTO)).thenReturn(expectedId);

        // Act
        BookDTO createdBookDTO = bookService.create(bookDTO);

        // Assert
        assertEquals(expectedId, createdBookDTO.getId());
        // Additional assertions can be made to verify that the bookDTO is returned as expected
    }

    @Test
    public void testGetTopAuthorsWithCharacterCount() {
        // Arrange
        BookRepositoryImpl bookRepository = Mockito.mock(BookRepositoryImpl.class);
        BookServiceImpl bookService = new BookServiceImpl(bookRepository);
        List<BookDTO> allBooks = new ArrayList<>();
        // Add some dummy data to allBooks

        // Mocking bookRepository behavior for the method call
        when(bookRepository.findAll()).thenReturn(allBooks);

        // Act
        List<Object[]> topAuthorsWithCharacterCount = bookService.getTopAuthorsWithCharacterCount("a");

        // Assert
        // You can add assertions to verify the filtering, sorting, and limiting logic
    }
}
